solver: move ops to separate files
Preparation for making these unit testable.
